Initial commit for simulation clock
Working test of a possible implementation of a simulation clock for simulation
engines using fixed size frames. This is instead of using doubles to hold and
step time. Using doubles will not be type safe and allow for stepping
backwards as well as not keeping to the stipulated frame rate. Double will be
much faster though.